+++ checker/src/main/java/org/checkerframework/checker/lock/LockVisitor.java
@@ -542,14 +542,14 @@ public class LockVisitor extends BaseTypeVisitor<LockAnnotatedTypeFactory> {
     }
 
     if (enclosingMethodElement != null) {
-      SideEffectAnnotation seaOfContainingMethod =
+      SideEffectAnnotation seaOfEnclosingMethod =
           atypeFactory.methodSideEffectAnnotation(enclosingMethodElement, false);
 
-      if (seaOfInvokedMethod.isWeakerThan(seaOfContainingMethod)) {
+      if (seaOfInvokedMethod.isWeakerThan(seaOfEnclosingMethod)) {
         checker.reportError(
             methodInvocationTree,
             "method.guarantee.violated",
-            seaOfContainingMethod.getNameOfSideEffectAnnotation(),
+            seaOfEnclosingMethod.getNameOfSideEffectAnnotation(),
             enclosingMethodElement.getSimpleName(),
             methodElement.getSimpleName(),
             seaOfInvokedMethod.getNameOfSideEffectAnnotation());
@@ -807,12 +807,11 @@ public class LockVisitor extends BaseTypeVisitor<LockAnnotatedTypeFactory> {
     if (enclosingMethod != null) {
       methodElement = TreeUtils.elementFromDeclaration(enclosingMethod);
 
-      SideEffectAnnotation seaOfContainingMethod =
+      SideEffectAnnotation seaOfEnclosingMethod =
           atypeFactory.methodSideEffectAnnotation(methodElement, false);
 
-      if (!seaOfContainingMethod.isWeakerThan(SideEffectAnnotation.LOCKINGFREE)) {
-        checker.reportError(
-            tree, "synchronized.block.in.lockingfree.method", seaOfContainingMethod);
+      if (!seaOfEnclosingMethod.isWeakerThan(SideEffectAnnotation.LOCKINGFREE)) {
+        checker.reportError(tree, "synchronized.block.in.lockingfree.method", seaOfEnclosingMethod);
       }
     }