+++ framework/src/main/java/org/checkerframework/framework/util/CheckerMain.java
@@ -689,7 +689,7 @@ public class CheckerMain {
     }
     String name = cls.getName();
     String classFileName;
-    /* name is something like pakkage.name.ContainingClass$ClassName. We need to turn this into ContainingClass$ClassName.class. */
+    /* name is something like pakkage.name.EnclosingClass$ClassName. We need to turn this into EnclosingClass$ClassName.class. */
     {
       int idx = name.lastIndexOf('.');
       classFileName = (idx == -1 ? name : name.substring(idx + 1)) + ".class";